requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
langchain==0.1.16
# openai==0.28.0
openai==1.26.0
python_dotenv==1.0.1
tiktoken==0.6.0

docxtpl==0.17.0
python-docx==1.1.2